requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
PyPDF2
langchain
langchain-google-genai
langchain_community
faiss-cpu
google-generativeai
streamlit
matplotlib
networkx